fix: dtype mismatch in qualitative_eval.py's generate() helper
etext was force-cast to .float() (matching CTCLIPStage2VLM.forward()'s
fp32-only training path), but the whole model is cast to bfloat16 for
inference here — the aggregator's FiLM/attention-conditioning Linears
(bfloat16 weights) then received a float32 etext, crashing with "mat1 and
mat2 must have the same dtype". Cast etext to model.llm.dtype instead.

Verified via job 66646634 (checkpoint_llama3b_v2, val_loss=0.4225): mean
cosine(instruction A, instruction B) = 0.4729 across 5 scans, all
individually below the 0.85 divergence threshold — "CONDITIONING WORKING".

+++ scripts/qualitative_eval.py
@@ -114,7 +114,10 @@ def generate(model, features_1: torch.Tensor, instruction: str, max_new_tokens: 
     -> llm.generate()) without modifying forward() itself. Greedy decoding
     (do_sample=False, num_beams=1) for reproducibility."""
     device = features_1.device
-    etext = model.instruction_encoder([instruction]).float()
+    # The whole model (including the aggregator's FiLM/attention-conditioning
+    # Linears) was cast to bfloat16 for inference — etext must match, not the
+    # float32 CTCLIPStage2VLM.forward() uses in its (fp32-only) training path.
+    etext = model.instruction_encoder([instruction]).to(model.llm.dtype)
     visual = model.projector(features_1, etext)
     visual = model.visual_proj(visual).to(model.llm.dtype)
